Tentative fix for edvin/tornadofx/issues/271
As described in https://github.com/edvin/tornadofx/issues/271, the `text` property('s value) on `AbstractField` can sometimes be `null`. This merely makes that explicit and replaces the null-unsafe `text.isNotEmpty()` with the null-safe `!text.isNullOrBlank()`.

I do not have the wherewithal to test this at this time, so request that testing is done upstream before merging this to make sure this doesn't break anything (as unlikely as that is) and to determine if this issue is present elsewhere.