Make Styleable.style(...) more general by using reflection
Rewrote Styleable.style(...) to use reflection to set the style
property instead of checking if it's one of the known classes that
contain the style property.

Now it should also work with unknown classes that also implement the
Styleable interface together with a setStyle(...) method.

Also removed TableColumnBase<*, *>.style(...) as it was redundant.