Make Styleable.style(...) more general by using reflection
Rewrote Styleable.style(...) to use reflection to set the style
property instead of checking if it's one of the known classes that
contain the style property.

Now it should also work with unknown classes that also implement the
Styleable interface together with a setStyle(...) method.

Also removed TableColumnBase<*, *>.style(...) as it was redundant.

Update the setStyle(...) method invocation for Java 9

Wrapped the Styleable.setStyle(...) method invocation in a try-catch
clause so that when used on an non-public object, it will gracefully
fail in Java 9.

Updated the inlineStyle() function to confirm the previously supported
types still work as before.

Updated the inlineStyle() test function

Updated the inlineStyle() test function to confirm the previously
supported types still work as before.

Make Styleable.style(...) more general by using reflection

Rewrote Styleable.style(...) to use reflection to set the style
property instead of checking if it's one of the known classes that
contain the style property.

Now it should also work with unknown classes that also implement the
Styleable interface together with a setStyle(...) method.

Also removed TableColumnBase<*, *>.style(...) as it was redundant.

Make Styleable.style(...) more general by using reflection

Rewrote Styleable.style(...) to use reflection to set the style
property instead of checking if it's one of the known classes that
contain the style property.

Now it should also work with unknown classes that also implement the
Styleable interface together with a setStyle(...) method.

Also removed TableColumnBase<*, *>.style(...) as it was redundant.